kernel: support >2GB malloc on 64-bit, add malloc4g smoke test
- trunkmem block size/used now share one address-width word instead of
  a uint32_t bitfield, removing the 2GB single-allocation cap
- IRQ/interrupt paths updated for the large demand-paged allocations
- malloc4g test app: 5GB + 1GB malloc, touch every page, verify content

+++ kernel/kernel/src/svc.c
@@ -143,7 +143,7 @@ static void sys_usleep(context_t* ctx, uint32_t count) {
     proc_usleep(ctx, count);
 }
 
-static ewokos_addr_t sys_malloc(int32_t size) {
+static ewokos_addr_t sys_malloc(int64_t size) {
     return (ewokos_addr_t)proc_malloc(get_current_proc(), size);
 }
